requirements.txt
flask==3.0.0
transformers==4.36.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
PyMuPDF==1.23.8
openai==1.3.0
torch==2.1.0 